+++ R/ncvsurv.R
@@ -27,7 +27,7 @@ ncvsurv <- function(X, y, penalty=c("MCP", "SCAD", "lasso"), gamma=switch(penalt
 
   ## Set up XX, yy, lambda
   tOrder <- order(y[,1])
-  yy <- as.numeric(y[tOrder,1])
+  yy <- as.double(y[tOrder,1])
   Delta <- y[tOrder,2]
   n <- length(yy)
   XX <- std(X[tOrder,,drop=FALSE])
